pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "flask>=3.1.2",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "flask-session>=0.8.0",
    "oauthlib>=3.3.1",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
]
